requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
attrs==26.1.0
certifi==2026.2.25
chardet==5.2.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.0
cuda-toolkit==13.0.2
cycler==0.12.1
DataProperty==1.1.0
datasets==4.8.4
dill==0.4.1
einops==0.8.2
evaluate==0.4.6
filelock==3.25.2
flash_attn==2.8.3
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gurobipy==13.0.1
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.8.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
kiwisolver==1.5.0
lm_eval==0.4.11
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mbstrdecoder==1.1.4
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.6.1
nltk==3.9.4
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti==13.0.85
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver==12.0.4.66
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse==12.6.3.3
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu12==2.21.5
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==3.0.1
pathvalidate==3.3.1
peft==0.18.1
pillow==12.1.1
pip==26.0.1
portalocker==3.2.0
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
Pygments==2.19.2
pymetis==2025.2.2
pyparsing==3.3.2
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.2.28
requests==2.33.0
rich==14.3.3
rouge_score==0.1.2
sacrebleu==2.6.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
sqlitedict==2.1.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.10.0
tcolorpy==0.1.7
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.3
transformers==5.3.0
triton==3.1.0
typepy==1.3.4
typer==0.24.1
typing_extensions==4.15.0
urllib3==2.6.3
wheel==0.46.3
word2number==1.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
